refa: name duplicate check move to blackjackGame
